Move core/utils.go to separate package and split into multiple files
